requirements.txt
colorama==0.4.6
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pandas
pyodbc
python-decouple
